rename all component files back to original name
