+++ src/components/qr-generator.tsx
